requirements.txt
# DexIQ AI Presenter - Python Dependencies

# Backend framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0

# Agent orchestration
langgraph>=0.2.0
langchain-openai>=0.2.0
langchain-core>=0.3.0

# Control interface
chainlit>=1.0.0

# HTTP client (for ElevenLabs API)
httpx>=0.25.0

# Configuration & validation
pyyaml>=6.0
pydantic>=2.0
pydantic-settings>=2.0

# Environment
python-dotenv>=1.0

# QR code generation
qrcode[pil]>=7.4

# Audio utilities
pydub>=0.25.1

# Testing
pytest>=7.0
pytest-asyncio>=0.21.0
